requirements.txt
tensorflow
streamlit
numpy
pillow
scikit-learn